P2Pの動作確認済み

+++ Core/Inc/FlightData/AutopilotData.hpp
@@ -5,10 +5,10 @@
 
 // Autopilot data populated from P2P packets (stored as int32_t per packet spec)
 struct AutopilotData {
-    int8_t pitch = 0;     // mapped parameter, -127..127
-    int8_t roll = 0;      // mapped parameter, -127..127
-    int8_t yaw = 0;       // mapped parameter, -127..127
-    uint8_t throttle = 0;  // mapped parameter, 0..255 (used as altitude proxy; 255 -> 1.0 m)
+    int32_t pitch = 0;     // mapped parameter, -127..127
+    int32_t roll = 0;      // mapped parameter, -127..127
+    int32_t yaw = 0;       // mapped parameter, -127..127
+    int32_t throttle = 0;  // mapped parameter, 0..255 (used as altitude proxy; 255 -> 1.0 m)
 
 };
